backend/requirements.txt
# Contenido de backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
psycopg2-binary
Pillow
anthropic>=0.25.0
python-multipart